+++ src/test/java/lotto/LottoServiceTest.java
@@ -1,7 +1,9 @@
-package lottogame;
+package lotto;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
+import lotto.domain.LottoService;
+import lotto.domain.Payment;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -10,7 +12,7 @@ public class LottoServiceTest {
     @DisplayName("로또 서비스 객체 생성")
     @Test
     void 로또_서비스_객체_생성() {
-        Payment payment = new Payment("3000");
+        String payment = "3000";
 
         assertDoesNotThrow(() -> new LottoService(payment));
     }
