+++ src/test/java/lotto/WinningNumberTest.java
@@ -1,8 +1,9 @@
-package lottogame;
+package lotto;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
+import lotto.domain.WinningNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
